refactor(lecture): 패지키 구조 변경
